pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "engram-memory"
version = "1.0.0"
description = "Cognitive memory architecture for persistent AI agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "invisiblemonsters"}]
keywords = ["ai", "memory", "agents", "vector-search", "lancedb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "lancedb>=0.6",
    "pyarrow>=14.0",
    "requests>=2.28",
]

[project.optional-dependencies]
embeddings = ["sentence-transformers>=2.2"]
identity = ["pynacl>=1.5"]
openai = ["openai>=1.0"]
config = ["pyyaml>=6.0"]
all = ["sentence-transformers>=2.2", "pynacl>=1.5", "openai>=1.0", "pyyaml>=6.0"]
dev = ["pytest>=7.0", "sentence-transformers>=2.2", "pynacl>=1.5", "pyyaml>=6.0"]

[project.urls]
Homepage = "https://github.com/invisiblemonsters/engram"
Repository = "https://github.com/invisiblemonsters/engram"

[tool.setuptools.packages.find]
include = ["engram*"]
exclude = ["engram_core*", "tests*", "examples*"]
